+++ ICD.Connect.Devices/Controls/IDeviceControl.cs
@@ -7,6 +7,8 @@ using ICD.Common.Utils.Extensions;
 using ICD.Connect.API.Attributes;
 using ICD.Connect.API.Nodes;
 using ICD.Connect.Devices.Proxies.Controls;
+using ICD.Connect.Telemetry;
+using ICD.Connect.Telemetry.Attributes;
 
 namespace ICD.Connect.Devices.Controls
 {
@@ -14,7 +16,7 @@ namespace ICD.Connect.Devices.Controls
 	/// IDeviceControl contains the usage features for a parent device.
 	/// </summary>
 	[ApiClass(typeof(ProxyDeviceControl))]
-	public interface IDeviceControl : IConsoleNode, IStateDisposable
+	public interface IDeviceControl : IConsoleNode, IStateDisposable, ITelemetryProvider
 	{
 		/// <summary>
 		/// Gets the parent device for this control.
@@ -25,12 +27,14 @@ namespace ICD.Connect.Devices.Controls
 		/// Gets the id for this control.
 		/// </summary>
 		[ApiProperty(DeviceControlApi.PROPERTY_ID, DeviceControlApi.HELP_PROPERTY_ID)]
+		[StaticPropertyTelemetry(ControlTelemetryNames.ID)]
 		int Id { get; }
 
 		/// <summary>
 		/// Gets the human readable name for this control.
 		/// </summary>
 		[ApiProperty(DeviceControlApi.PROPERTY_NAME, DeviceControlApi.HELP_PROPERTY_NAME)]
+		[StaticPropertyTelemetry(ControlTelemetryNames.NAME)]
 		string Name { get; }
 
 		/// <summary>