+++ ICD.Connect.Devices/DeviceTelemetryNames.cs
@@ -1,0 +1,13 @@
+﻿namespace ICD.Connect.Devices
+{
+	public static class DeviceTelemetryNames
+	{
+		public const string POWER_STATE = "IsPowered";
+		public const string POWER_ON = "PowerOn";
+		public const string POWER_OFF = "PowerOff";
+		public const string POWER_STATE_CHANGED = "OnPowerStateChanged";
+
+		public const string ONLINE_STATE = "IsOnline";
+		public const string ONLINE_STATE_CHANGED = "OnIsOnlineChanged";
+	}
+}