+++ CHANGELOG.md
@@ -10,6 +10,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
  - Originators and settings moved into subdirectories
 
+## [6.4.0] - 2019-05-15
+### Added
+ - Added telemetry features to base devices
+ 
+### Changed
+ - Clearer exception when control lookup fails
+
 ## [6.3.2] - 2019-02-07
 ### Changed
  - Better error logging for bad control lookups